Moving away from lat long coords

+++ Runtime/Interaction/HPUIInteractorRayAngle.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using UnityEngine;
 using System;
 
@@ -7,21 +6,22 @@ namespace ubco.ovilab.HPUI.Interaction
     [Serializable]
     public struct HPUIInteractorRayAngle
     {
-        public int x, z;
+        public float x, z;
 
-        public HPUIInteractorRayAngle(int x, int z)
+        public HPUIInteractorRayAngle(float x, float z)
         {
             this.x = x;
             this.z = z;
         }
 
-        public static Vector3 GetDirection(int x, int z, Vector3 right, Vector3 forward, Vector3 up, bool flipZAngles)
+        public static Vector3 GetDirection(float x, float z, Vector3 right, Vector3 forward, Vector3 up, bool flipZAngles)
         {
-            int x_ = x,
-                z_ = flipZAngles ? -z : z;
+            float x_ = x,
+                  z_ = flipZAngles ? -z : z;
 
             Quaternion rotation = Quaternion.AngleAxis(x_, right) * Quaternion.AngleAxis(z_, forward);
             return rotation * up;
+            // return 
         }
 
         public Vector3 GetDirection(Transform attachTransform, bool flipZAngles)
